Move PowerQueryParser to a separate file
